Crate bestia_dev_new_rust_wasm_pwa

bestia_dev_new_rust_wasm_pwa lib.rs is just for the wasm_bindgen_start function and to connect to all the modules. and for the big doc comments

Modules

dom_mod 🔒

dom_mod.rs manipulates the html dom

prepare_zip_mod 🔒

prepare_zip_mod.rs should not use any javascript objects. Pure Rust. All the javascript objects, functions and conversion should be in web_sys_mod.rs.

web_sys_mod 🔒

web_sys_mod.rs helper functions for web_sys, window, document, dom, console, local_storage, session_storage,… Trying to isolate/hide all javascript code and conversion here.

Structs

PwaData

Constants

__wasm_bindgen_generated_wasm_bindgen_start__const 🔒

Functions

wasm_bindgen_start

To start the Wasm application, wasm_bindgen runs this functions